+++ client/src/pages/dashboard/users/students/index.tsx
@@ -1,6 +1,7 @@
 import Link from "next/link";
 import { useRouter } from "next/router";
 import React, { Suspense, useEffect, useState } from "react";
+import { toast } from "sonner";
 
 import { ProtectedPage } from "@/components/layout";
 import { Button } from "@/components/ui/button";
@@ -24,6 +25,16 @@ export default function PageConfig() {
     </ProtectedPage>
   );
 }
+
+// For CSV export
+type StoredRecord = {
+  studentId: string;
+  firstName: string;
+  lastName: string;
+  password: string;
+  yearLevel: number | string;
+  schoolName: string;
+};
 
 function UserList() {
   const router = useRouter();
@@ -60,6 +71,66 @@ function UserList() {
 
   if (error) return <div>Error: {error.message}</div>;
 
+  /**
+   * Download a CSV file from the db data.
+   * The CSV will contain the following columns:
+   *   Student ID, First Name, Last Name, Year Level, School Name,
+   */
+  const downloadStudentsCSV = () => {
+    // instead of getting from localStorage, use the data fetched from the API
+    const storedData: StoredRecord[] = (data ?? [])
+      .filter((record) => record.student_id !== undefined)
+      .map((record) => ({
+        studentId: record.student_id as string,
+        firstName: record.first_name,
+        lastName: record.last_name,
+        password: record.plaintext_password ?? "N/A",
+        yearLevel: record.year_level,
+        schoolName: record.school.name,
+      }));
+
+    if (storedData.length === 0) {
+      toast.warning("No data available to export.");
+      return;
+    }
+
+    // CSV headers in the order you want them
+    const csvHeaders = [
+      "Student ID",
+      "First Name",
+      "Last Name",
+      "Password",
+      "Year Level",
+      "School Name",
+    ];
+
+    // Build each row from storedData
+    const csvRows = storedData.map((record) => [
+      record.studentId,
+      record.firstName,
+      record.lastName,
+      record.password,
+      record.yearLevel,
+      record.schoolName,
+    ]);
+
+    // Convert to CSV string
+    const csvContent = [csvHeaders, ...csvRows]
+      .map((row) => row.map((value) => `"${value}"`).join(","))
+      .join("\n");
+
+    // Trigger a download
+    const blob = new Blob([csvContent], { type: "text/csv;charset=utf-8;" });
+    const url = URL.createObjectURL(blob);
+    const link = document.createElement("a");
+
+    link.href = url;
+    link.setAttribute("download", "Student_Data_WAJO.csv");
+    document.body.appendChild(link);
+    link.click();
+    document.body.removeChild(link);
+  };
+
   return (
     <div className="m-4 space-y-4">
       <div className="flex justify-between">
@@ -77,6 +148,16 @@ function UserList() {
       </div>
 
       <Suspense>
+        <div className="flex justify-end">
+          <Button
+            variant="outline"
+            className="h-auto"
+            onClick={downloadStudentsCSV}
+            disabled={isLoading}
+          >
+            Download this page as CSV
+          </Button>
+        </div>
         <div>
           <DataGrid
             datacontext={data ?? []}
